% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-pk-advan.R
\name{.ad_sqrt}
\alias{.ad_sqrt}
\title{PKADVAN-style analytical step functions (R port of src/nm_pk_pkadvan.h)}
\usage{
.ad_sqrt(x)
}
\description{
PKADVAN-style analytical step functions (R port of src/nm_pk_pkadvan.h)
}
\keyword{internal}
