% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-bayes.R
\name{.nm_bayes_prior_spec}
\alias{.nm_bayes_prior_spec}
\title{Build prior specification vectors for C++ Bayesian MCMC}
\usage{
.nm_bayes_prior_spec(model, par0, prior = NULL)
}
\description{
Build prior specification vectors for C++ Bayesian MCMC
}
\keyword{internal}
