% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-nonmem-bench.R
\name{.nm_bench_data_from_simtab}
\alias{.nm_bench_data_from_simtab}
\title{Build estimation dataset from NONMEM simulation table}
\usage{
.nm_bench_data_from_simtab(design, nm_tab, input_cols)
}
\description{
Build estimation dataset from NONMEM simulation table
}
\keyword{internal}
