% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-covariates.R
\name{.nm_bind_covariates}
\alias{.nm_bind_covariates}
\title{Bind covariate columns into a PRED evaluation environment}
\usage{
.nm_bind_covariates(model, subj, env)
}
\arguments{
\item{model}{An \code{nm_model} object.}

\item{subj}{Subject event data (one ID).}

\item{env}{Environment to augment (typically from \code{.nm_make_env}).}
}
\description{
Bind covariate columns into a PRED evaluation environment
}
\keyword{internal}
