% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-blq.R
\name{.nm_blq_active}
\alias{.nm_blq_active}
\title{Is the BLQ (M3/M4) likelihood active for a config?}
\usage{
.nm_blq_active(cfg)
}
\description{
Is the BLQ (M3/M4) likelihood active for a config?
}
\keyword{internal}
