% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-blq.R
\name{.nm_blq_phi}
\alias{.nm_blq_phi}
\title{Standard-normal CDF clamped away from 0/1 for stable logs.}
\usage{
.nm_blq_phi(z)
}
\description{
Standard-normal CDF clamped away from 0/1 for stable logs.
}
\keyword{internal}
