% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{.nm_ctl_pk_scaling}
\alias{.nm_ctl_pk_scaling}
\title{Compartment scaling lines for $PK (NONMEM Sx convention: F = A(obs)/S(obs))}
\usage{
.nm_ctl_pk_scaling(advan, trans, ode_ncomp = 2L)
}
\description{
Compartment scaling lines for $PK (NONMEM Sx convention: F = A(obs)/S(obs))
}
\keyword{internal}
