% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{.nm_ctl_pk_scaling_ode}
\alias{.nm_ctl_pk_scaling_ode}
\title{ODE model S1..Sn scaling for $PK (supports NCOMPS up to 10)}
\usage{
.nm_ctl_pk_scaling_ode(ncomp = 2L, oral = TRUE, advan = 6L)
}
\description{
ODE model S1..Sn scaling for $PK (supports NCOMPS up to 10)
}
\keyword{internal}
