% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{.nm_ctl_pk_trans6_cardano_roots}
\alias{.nm_ctl_pk_trans6_cardano_roots}
\title{Parameter name aliases for NM-TRAN (macro models use Q not Q2, etc.)}
\usage{
.nm_ctl_pk_trans6_cardano_roots()
}
\description{
Parameter name aliases for NM-TRAN (macro models use Q not Q2, etc.)
}
\keyword{internal}
