% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{.nm_ctl_pk_volume_aliases}
\alias{.nm_ctl_pk_volume_aliases}
\title{Volume aliases required by NM-TRAN before Sx assignments (macro CL/VC/VP models)}
\usage{
.nm_ctl_pk_volume_aliases(advan, trans)
}
\description{
Volume aliases required by NM-TRAN before Sx assignments (macro CL/VC/VP models)
}
\keyword{internal}
