% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_default_input_col}
\alias{.nm_default_input_col}
\title{Default value when an $INPUT column is absent from the dataset}
\usage{
.nm_default_input_col(col)
}
\description{
Default value when an $INPUT column is absent from the dataset
}
\keyword{internal}
