% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-bayes.R
\name{.nm_est_bayes}
\alias{.nm_est_bayes}
\title{NONMEM-style full Bayesian MCMC (block Gibbs / Metropolis-Hastings)}
\usage{
.nm_est_bayes(
  model,
  data,
  par0,
  backend = "cpp",
  grad = "auto",
  pk_engine = "auto",
  engine = "auto",
  control = list(),
  n_burn = 100L,
  n_sample = 500L,
  n_thin = 1L,
  seed = 1L,
  step_scale = NULL,
  prior = NULL,
  store_eta_chain = TRUE,
  sampler = c("mh", "hmc", "nuts"),
  hmc_epsilon = 0.05,
  hmc_leap = 10L,
  nuts_depth = 5L
)
}
\description{
NONMEM-style full Bayesian MCMC (block Gibbs / Metropolis-Hastings)
}
\keyword{internal}
