% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est-methods.R
\name{.nm_est_foce}
\alias{.nm_est_foce}
\title{First-order conditional estimation (FOCE)}
\usage{
.nm_est_foce(
  model,
  data,
  par0,
  backend = "cpp",
  grad = "auto",
  pk_engine = "auto",
  engine = "auto",
  control = list(),
  max_outer = 20L,
  tol = 1e-04
)
}
\description{
First-order conditional estimation (FOCE)
}
\keyword{internal}
