% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei.R
\name{.nm_est_focei}
\alias{.nm_est_focei}
\title{FOCE with interaction (FOCEI)}
\usage{
.nm_est_focei(
  model,
  data,
  par0,
  backend = "cpp",
  grad = "auto",
  pk_engine = "auto",
  engine = "auto",
  control = list(),
  max_outer = 10L,
  tol = 1e-05,
  ...
)
}
\description{
FOCE with interaction (FOCEI)
}
\keyword{internal}
