% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-imp.R
\name{.nm_est_imp}
\alias{.nm_est_imp}
\title{Importance sampling (IMP) estimation}
\usage{
.nm_est_imp(
  model,
  data,
  par0,
  backend = "cpp",
  grad = "auto",
  pk_engine = "auto",
  engine = "auto",
  control = list(),
  n_imp = 50L,
  n_quad = 5L,
  ...
)
}
\description{
Uses Laplace eta-modes and Gaussian importance weights around each mode.
}
\keyword{internal}
