% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est-methods.R
\name{.nm_est_saem}
\alias{.nm_est_saem}
\title{Stochastic approximation EM (SAEM)}
\usage{
.nm_est_saem(
  model,
  data,
  par0,
  backend = "cpp",
  grad = "auto",
  pk_engine = "auto",
  engine = "auto",
  control = list(),
  n_iter = 100L,
  n_burn = 20L,
  n_mcmc = 1L,
  sa_rate = 0.6,
  seed = 1L
)
}
\description{
Stochastic approximation EM (SAEM)
}
\keyword{internal}
