% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_expand_addl}
\alias{.nm_expand_addl}
\title{Expand NONMEM ADDL/II implied doses into explicit dosing events.
ADDL is the number of additional doses; II (or TAU) is the interval.}
\usage{
.nm_expand_addl(subj)
}
\description{
Expand NONMEM ADDL/II implied doses into explicit dosing events.
ADDL is the number of additional doses; II (or TAU) is the interval.
}
\keyword{internal}
