% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est-methods.R
\name{.nm_fit_all_eta_modes}
\alias{.nm_fit_all_eta_modes}
\title{Fit post-hoc ETAs for all subjects (FOCE / FOCEI helper).}
\usage{
.nm_fit_all_eta_modes(
  model,
  dat,
  theta,
  omega,
  sigma,
  eta_mat = NULL,
  pk_engine = "auto",
  control = list()
)
}
\description{
Fit post-hoc ETAs for all subjects (FOCE / FOCEI helper).
}
\keyword{internal}
