% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-fo.R
\name{.nm_fo_omega_nll}
\alias{.nm_fo_omega_nll}
\title{FO omega term: \eqn{\sum_i \log|\Omega + G_i^{\top} R_i^{-1} G_i|}}
\usage{
.nm_fo_omega_nll(model, fo_cache, omega)
}
\description{
FO omega term: \eqn{\sum_i \log|\Omega + G_i^{\top} R_i^{-1} G_i|}
}
\keyword{internal}
