% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-fo.R
\name{.nm_fo_subject_marginal_r}
\alias{.nm_fo_subject_marginal_r}
\title{Per-subject FO marginal contribution: \eqn{\log|V| + r^{\top} V^{-1} r}.}
\usage{
.nm_fo_subject_marginal_r(f0, dv, G, R_diag, OM)
}
\description{
Per-subject FO marginal contribution: \eqn{\log|V| + r^{\top} V^{-1} r}.
}
\keyword{internal}
