% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei-sens.R
\name{.nm_focei_needs_fd_g}
\alias{.nm_focei_needs_fd_g}
\title{Models where eta sensitivities are discontinuous or AD-unreliable.}
\usage{
.nm_focei_needs_fd_g(model, subj = NULL)
}
\description{
Models where eta sensitivities are discontinuous or AD-unreliable.
}
\keyword{internal}
