% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei.R
\name{.nm_focei_nll_internal}
\alias{.nm_focei_nll_internal}
\title{Population FOCEI -2LL (C++ when numeric, R when AD active).}
\usage{
.nm_focei_nll_internal(
  model,
  data,
  theta,
  omega,
  sigma,
  eta_mat,
  pk_engine = "auto"
)
}
\description{
Population FOCEI -2LL (C++ when numeric, R when AD active).
}
\keyword{internal}
