% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei.R
\name{.nm_focei_omega_matrix}
\alias{.nm_focei_omega_matrix}
\title{Build omega matrix for FOCEI (numeric or AD list-matrix).}
\usage{
.nm_focei_omega_matrix(omega, n_eta, omega_struct = "diag")
}
\description{
Build omega matrix for FOCEI (numeric or AD list-matrix).
}
\keyword{internal}
