% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei.R
\name{.nm_focei_residual_var}
\alias{.nm_focei_residual_var}
\title{FOCE residual variance R_j (no eta variance in V).}
\usage{
.nm_focei_residual_var(f, sigma, error = "propadd")
}
\description{
FOCE residual variance R_j (no eta variance in V).
}
\keyword{internal}
