% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei-setup.R
\name{.nm_focei_sensitivity_grad}
\alias{.nm_focei_sensitivity_grad}
\title{Gradient of FOCEI at fixed conditional eta modes (total grad when eta is at mode).}
\usage{
.nm_focei_sensitivity_grad(model, data, par, eta_mat, pk_engine)
}
\description{
Gradient of FOCEI at fixed conditional eta modes (total grad when eta is at mode).
}
\keyword{internal}
