% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei-sens.R
\name{.nm_focei_shi_eps}
\alias{.nm_focei_shi_eps}
\title{Shi-style adaptive FD step for FOCEI G (eta or F scale).}
\usage{
.nm_focei_shi_eps(x, f_ref = NULL)
}
\description{
Shi-style adaptive FD step for FOCEI G (eta or F scale).
}
\keyword{internal}
