% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei-sens.R
\name{.nm_focei_subject_G_fd}
\alias{.nm_focei_subject_G_fd}
\title{Central FD with fixed or Shi step sizes.}
\usage{
.nm_focei_subject_G_fd(
  model,
  subj,
  theta,
  omega,
  eta,
  sigma,
  pk_engine,
  f,
  n_obs,
  n_eta,
  use_shi = TRUE
)
}
\description{
Central FD with fixed or Shi step sizes.
}
\keyword{internal}
