% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-focei.R
\name{.nm_focei_subject_nll_internal}
\alias{.nm_focei_subject_nll_internal}
\title{Per-subject FOCEI -2LL at a fixed eta mode (R path, AD-capable).}
\usage{
.nm_focei_subject_nll_internal(
  model,
  subj,
  theta,
  omega,
  sigma,
  eta,
  pk_engine = "auto",
  interaction = TRUE
)
}
\description{
Per-subject FOCEI -2LL at a fixed eta mode (R path, AD-capable).
}
\keyword{internal}
