% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-imp.R
\name{.nm_imp_emit_diagnostics}
\alias{.nm_imp_emit_diagnostics}
\title{Emit one-time IMP diagnostics (proposal downgrade / low ESS).}
\usage{
.nm_imp_emit_diagnostics(diag_env, n_imp)
}
\description{
Emit one-time IMP diagnostics (proposal downgrade / low ESS).
}
\keyword{internal}
