% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-imp.R
\name{.nm_imp_num_hessian}
\alias{.nm_imp_num_hessian}
\title{Central-difference Hessian for small eta dimension.}
\usage{
.nm_imp_num_hessian(fn, x, h = 1e-04)
}
\description{
Central-difference Hessian for small eta dimension.
}
\keyword{internal}
