% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_input_event_vectors}
\alias{.nm_input_event_vectors}
\title{Row-level S/F vectors for C++ PK (data override when user supplied columns)}
\usage{
.nm_input_event_vectors(subj)
}
\description{
Row-level S/F vectors for C++ PK (data override when user supplied columns)
}
\keyword{internal}
