% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_ipred_align}
\alias{.nm_ipred_align}
\title{Map ipred from an expanded event table back to the original subject rows.}
\usage{
.nm_ipred_align(subj, subj_ev, ipred_ev)
}
\description{
Map ipred from an expanded event table back to the original subject rows.
}
\keyword{internal}
