% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-crypto.R
\name{.nm_job_read_args}
\alias{.nm_job_read_args}
\title{Read job args, decrypting args.enc when a DEK is present.}
\usage{
.nm_job_read_args(job_path)
}
\description{
Read job args, decrypting args.enc when a DEK is present.
}
\keyword{internal}
