% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-blq.R
\name{.nm_model_has_covariates}
\alias{.nm_model_has_covariates}
\title{Does a model declare covariates?}
\usage{
.nm_model_has_covariates(model)
}
\description{
Does a model declare covariates?
}
\keyword{internal}
