% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-parse.R
\name{.nm_nm_expr_normalize}
\alias{.nm_nm_expr_normalize}
\title{Normalize NONMEM-style power (**) to R/C++ power (^) in a single expression line}
\usage{
.nm_nm_expr_normalize(line)
}
\description{
Normalize NONMEM-style power (**) to R/C++ power (^) in a single expression line
}
\keyword{internal}
