% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_normalize_dose_records}
\alias{.nm_normalize_dose_records}
\title{Normalize NONMEM RATE=-1 (bolus) and RATE=-2 (duration in AMT) dose rows}
\usage{
.nm_normalize_dose_records(dat)
}
\description{
Normalize NONMEM RATE=-1 (bolus) and RATE=-2 (duration in AMT) dose rows
}
\keyword{internal}
