% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_pred_f_cols}
\alias{.nm_pred_f_cols}
\title{Fx symbols assigned in $PK (exclude from data defaults when set in model)}
\usage{
.nm_pred_f_cols(model)
}
\description{
Fx symbols assigned in $PK (exclude from data defaults when set in model)
}
\keyword{internal}
