% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_pred_scale_cols}
\alias{.nm_pred_scale_cols}
\title{Sx symbols assigned in $PK (exclude from data defaults when set in model)}
\usage{
.nm_pred_scale_cols(model)
}
\description{
Sx symbols assigned in $PK (exclude from data defaults when set in model)
}
\keyword{internal}
