% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-profile-likelihood.R
\name{.nm_profile_ll_ci}
\alias{.nm_profile_ll_ci}
\title{Linear-interpolation likelihood CI from a profile.}
\usage{
.nm_profile_ll_ci(grid, ofv, est, target)
}
\description{
Linear-interpolation likelihood CI from a profile.
}
\keyword{internal}
