% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-remote-server.R
\name{.nm_remote_config_path}
\alias{.nm_remote_config_path}
\title{Config file for remote scheduler servers}
\usage{
.nm_remote_config_path()
}
\description{
Config file for remote scheduler servers
}
\keyword{internal}
