% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job-remote.R
\name{.nm_remote_job_cancel}
\alias{.nm_remote_job_cancel}
\title{Cancel a remote job}
\usage{
.nm_remote_job_cancel(server, job_id)
}
\arguments{
\item{server}{Server id.}

\item{job_id}{Remote job id.}
}
\description{
Cancel a remote job
}
\keyword{internal}
