% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job-remote.R
\name{.nm_remote_job_log}
\alias{.nm_remote_job_log}
\title{Remote worker log}
\usage{
.nm_remote_job_log(server, job_id, tail = 100L)
}
\arguments{
\item{server}{Server id.}

\item{job_id}{Remote job id.}

\item{tail}{Number of log lines.}
}
\description{
Remote worker log
}
\keyword{internal}
