% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job-remote.R
\name{.nm_remote_job_result}
\alias{.nm_remote_job_result}
\title{Fetch a remote job result}
\usage{
.nm_remote_job_result(server, job_id)
}
\arguments{
\item{server}{Server id.}

\item{job_id}{Remote job id.}
}
\description{
Fetch a remote job result
}
\keyword{internal}
