% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job-remote.R
\name{.nm_remote_job_status}
\alias{.nm_remote_job_status}
\title{Remote job status}
\usage{
.nm_remote_job_status(server, job_id)
}
\arguments{
\item{server}{Server id.}

\item{job_id}{Remote job id.}
}
\value{
Job metadata list.
}
\description{
Remote job status
}
\keyword{internal}
