% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-lik-config.R
\name{.nm_residual_var}
\alias{.nm_residual_var}
\title{Residual variance at observations (R path)}
\usage{
.nm_residual_var(f, sigma, error = "propadd", dvid = NULL, i = NULL)
}
\description{
Residual variance at observations (R path)
}
\keyword{internal}
