% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-run-info.R
\name{.nm_run_info_format_duration}
\alias{.nm_run_info_format_duration}
\title{Format run duration for display}
\usage{
.nm_run_info_format_duration(seconds)
}
\arguments{
\item{seconds}{Duration in seconds.}
}
\value{
Character label.
}
\description{
Format run duration for display
}
\keyword{internal}
