% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_scale_event_vectors}
\alias{.nm_scale_event_vectors}
\title{Row-level S vectors for C++ PK (data override when user supplied S columns)}
\usage{
.nm_scale_event_vectors(subj)
}
\description{
Row-level S vectors for C++ PK (data override when user supplied S columns)
}
\keyword{internal}
