% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{.nm_sim_combine_replicates}
\alias{.nm_sim_combine_replicates}
\title{Combine simulated replicates into one dataset with a REP column}
\usage{
.nm_sim_combine_replicates(sims)
}
\arguments{
\item{sims}{List of \code{nm_dataset} objects from \code{\link{nm_simulate}}.}
}
\value{
A single \code{nm_dataset} with all rows stacked.
}
\description{
Combine simulated replicates into one dataset with a REP column
}
\keyword{internal}
