% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{.nm_sim_pack_output}
\alias{.nm_sim_pack_output}
\title{Package simulation output for workspace persistence}
\usage{
.nm_sim_pack_output(sims)
}
\description{
Package simulation output for workspace persistence
}
\keyword{internal}
