% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{.nm_sim_parse_dose_table}
\alias{.nm_sim_parse_dose_table}
\title{Parse dose amount lines (\code{"TIME AMT"} or \code{"AMT"})}
\usage{
.nm_sim_parse_dose_table(text, default_time = 0)
}
\description{
Parse dose amount lines (\code{"TIME AMT"} or \code{"AMT"})
}
\keyword{internal}
