% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{.nm_sim_resize_subjects}
\alias{.nm_sim_resize_subjects}
\title{Resample or expand subjects in a template dataset}
\usage{
.nm_sim_resize_subjects(data, model, n_sub, seed = NULL)
}
\description{
Resample or expand subjects in a template dataset
}
\keyword{internal}
