% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{.nm_subject_events}
\alias{.nm_subject_events}
\title{Subject event table with ADDL/II doses expanded for PK simulation.}
\usage{
.nm_subject_events(subj)
}
\description{
Subject event table with ADDL/II doses expanded for PK simulation.
}
\keyword{internal}
