% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est-methods.R
\name{.nm_subject_laplace_nll}
\alias{.nm_subject_laplace_nll}
\title{Laplace approximation with Gaussian quadrature (NONMEM-style)}
\usage{
.nm_subject_laplace_nll(
  model,
  subj,
  theta,
  omega,
  sigma,
  gh,
  pk_engine = "auto"
)
}
\description{
Laplace approximation with Gaussian quadrature (NONMEM-style)
}
\keyword{internal}
