% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-params.R
\name{.nm_theta_default_bounds}
\alias{.nm_theta_default_bounds}
\title{Default THETA lower/upper from an initial estimate (0.1x and 10x).}
\usage{
.nm_theta_default_bounds(value)
}
\description{
Default THETA lower/upper from an initial estimate (0.1x and 10x).
}
\keyword{internal}
