% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{.nm_vpc_typical_pred}
\alias{.nm_vpc_typical_pred}
\title{Typical (population, eta = 0) prediction for each observation row.}
\usage{
.nm_vpc_typical_pred(model, data, theta, omega, sigma, pk_engine = "cpp")
}
\description{
Typical (population, eta = 0) prediction for each observation row.
}
\keyword{internal}
