% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{.nm_workspace_sim_dataset}
\alias{.nm_workspace_sim_dataset}
\title{Extract primary dataset from a loaded simulation object}
\usage{
.nm_workspace_sim_dataset(sim_obj)
}
\description{
Extract primary dataset from a loaded simulation object
}
\keyword{internal}
