% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-par-scale.R
\name{.nm_wrap_scaled_optim}
\alias{.nm_wrap_scaled_optim}
\title{Wrap optim objective/gradient for internal parameter scaling.}
\usage{
.nm_wrap_scaled_optim(model, par0, free, f_free, g_free, scale)
}
\description{
Wrap optim objective/gradient for internal parameter scaling.
}
\keyword{internal}
