% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{liberWorkbenchOutput}
\alias{liberWorkbenchOutput}
\title{Shiny output for the LibeR workbench}
\usage{
liberWorkbenchOutput(outputId, width = "100\%", height = "100vh")
}
\arguments{
\item{outputId}{Shiny output id.}

\item{width, height}{CSS dimensions.}
}
\description{
Shiny output for the LibeR workbench
}
