% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-nonmem-bench.R
\name{nm_bench_pairs}
\alias{nm_bench_pairs}
\title{All valid ADVAN/TRANS benchmark pairs}
\usage{
nm_bench_pairs()
}
\value{
Data frame with columns \code{advan}, \code{trans}, \code{tag}.
}
\description{
All valid ADVAN/TRANS benchmark pairs
}
\examples{
head(nm_bench_pairs())
}
