% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-nonmem-bench.R
\name{nm_bench_read_ext}
\alias{nm_bench_read_ext}
\title{Parse THETA/OMEGA/SIGMA and OBJ from a NONMEM .ext file}
\usage{
nm_bench_read_ext(ext_path)
}
\arguments{
\item{ext_path}{Path to \code{.ext} file.}
}
\value{
List with \code{theta}, \code{omega}, \code{sigma}, \code{obj}.
}
\description{
Parse THETA/OMEGA/SIGMA and OBJ from a NONMEM .ext file
}
\examples{
# Path to a NONMEM .ext file
}
