% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/engine.R
\name{nm_compile}
\alias{nm_compile}
\title{Compile a serializable model into a pointer-backed engine}
\usage{
nm_compile(model)
}
\arguments{
\item{model}{An [nm_model()].}
}
\description{
Compile a serializable model into a pointer-backed engine
}
