% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid-components.R
\name{nm_component_code}
\alias{nm_component_code}
\title{Generate restricted expression code for a model component}
\usage{
nm_component_code(component)
}
\arguments{
\item{component}{An [nm_component()].}
}
\description{
Generate restricted expression code for a model component
}
