% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_advan_choices}
\alias{nm_ctl_advan_choices}
\title{Valid ADVAN values for the GUI (standard PREDPP models)}
\usage{
nm_ctl_advan_choices()
}
\value{
Character vector of ADVAN numbers.
}
\description{
Valid ADVAN values for the GUI (standard PREDPP models)
}
\examples{
nm_ctl_advan_choices()
}
