% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_default_obscmp}
\alias{nm_ctl_default_obscmp}
\title{Default observation compartment for ADVAN/TRANS}
\usage{
nm_ctl_default_obscmp(advan, trans = NULL)
}
\description{
Default observation compartment for ADVAN/TRANS
}
\examples{
nm_ctl_default_obscmp(4L, 4L)
}
