% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_default_trans}
\alias{nm_ctl_default_trans}
\title{Default TRANS for an ADVAN (CL/V-style parameterization when available)}
\usage{
nm_ctl_default_trans(advan)
}
\description{
Default TRANS for an ADVAN (CL/V-style parameterization when available)
}
\examples{
nm_ctl_default_trans(4L)
}
