% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_error_highlight_symbols}
\alias{nm_ctl_error_highlight_symbols}
\title{Symbol lists for $ERROR syntax highlighting}
\usage{
nm_ctl_error_highlight_symbols()
}
\value{
List with \code{pk}, \code{flows}, \code{builtin}, and \code{variant}.
}
\description{
Symbol lists for $ERROR syntax highlighting
}
\examples{
nm_ctl_error_highlight_symbols()
}
