% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_essential_input_cols}
\alias{nm_ctl_essential_input_cols}
\title{Required dataset columns for \code{$INPUT} given ADVAN/TRANS}
\usage{
nm_ctl_essential_input_cols(advan = NULL, trans = NULL)
}
\arguments{
\item{advan}{ADVAN number; when \code{NULL} returns core identifiers only.}

\item{trans}{TRANS number.}
}
\value{
Character vector of required column names.
}
\description{
Required dataset columns for \code{$INPUT} given ADVAN/TRANS
}
\examples{
nm_ctl_essential_input_cols(4L, 4L)
}
