% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_is_valid_pair}
\alias{nm_ctl_is_valid_pair}
\title{Check whether an ADVAN/TRANS pair is valid in NONMEM}
\usage{
nm_ctl_is_valid_pair(advan, trans)
}
\description{
Check whether an ADVAN/TRANS pair is valid in NONMEM
}
\examples{
nm_ctl_is_valid_pair(4L, 4L)
}
