% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_output_extra_cols}
\alias{nm_ctl_output_extra_cols}
\title{Common derived columns offered for \code{$OUTPUT} selection}
\usage{
nm_ctl_output_extra_cols()
}
\value{
Character vector of column names.
}
\description{
Common derived columns offered for \code{$OUTPUT} selection
}
\examples{
nm_ctl_output_extra_cols()
}
