% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_picker_columns}
\alias{nm_ctl_picker_columns}
\title{All selectable names for the column picker (dataset + output extras)}
\usage{
nm_ctl_picker_columns(dataset_cols)
}
\arguments{
\item{dataset_cols}{Column names from the dataset file.}
}
\value{
Character vector.
}
\description{
All selectable names for the column picker (dataset + output extras)
}
\examples{
nm_ctl_picker_columns(c("ID", "TIME", "DV"))
}
