% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_subroutine_text}
\alias{nm_ctl_subroutine_text}
\title{$SUBROUTINE text for ODE ADVAN models}
\usage{
nm_ctl_subroutine_text(advan, trans)
}
\description{
$SUBROUTINE text for ODE ADVAN models
}
\examples{
nm_ctl_subroutine_text(4L, 4L)
}
