% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_trans_choices}
\alias{nm_ctl_trans_choices}
\title{Valid TRANS values for a given ADVAN (NONMEM PREDPP pairings)}
\usage{
nm_ctl_trans_choices(advan)
}
\arguments{
\item{advan}{ADVAN number.}
}
\value{
Character vector of valid TRANS values.
}
\description{
Valid TRANS values for a given ADVAN (NONMEM PREDPP pairings)
}
\examples{
nm_ctl_trans_choices(2L)
}
