% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-ctl-editor.R
\name{nm_ctl_use_ode}
\alias{nm_ctl_use_ode}
\title{Whether ADVAN uses the general ODE (\code{$SUBROUTINE}) path}
\usage{
nm_ctl_use_ode(advan)
}
\description{
Whether ADVAN uses the general ODE (\code{$SUBROUTINE}) path
}
\examples{
nm_ctl_use_ode(6L)
}
