% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-scm.R
\name{nm_forest_plot}
\alias{nm_forest_plot}
\title{Forest plot of retained SCM covariate effects}
\usage{
nm_forest_plot(scm, title = "SCM covariate effects")
}
\arguments{
\item{scm}{An \code{nm_scm} object from \code{\link{nm_scm}}.}

\item{title}{Plot title.}
}
\value{
A \code{ggplot} object (requires the suggested \pkg{ggplot2}).
}
\description{
Forest plot of retained SCM covariate effects
}
\examples{
\dontrun{
nm_forest_plot(nm_scm(model, data, candidates))
}
}
