% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_cancel}
\alias{nm_job_cancel}
\title{Cancel a running job}
\usage{
nm_job_cancel(job_id, job_root = nm_job_root())
}
\arguments{
\item{job_id}{Job identifier.}

\item{job_root}{Job storage directory.}
}
\value{
Updated status list.
}
\description{
Cancel a running job
}
\examples{
# After nm_job_submit(), use nm_job_cancel(job$id)
}
