% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_cleanup}
\alias{nm_job_cleanup}
\title{Remove finished jobs from disk}
\usage{
nm_job_cleanup(
  job_root = nm_job_root(),
  status = c("success", "error", "cancelled")
)
}
\arguments{
\item{job_root}{Job storage directory.}

\item{status}{Status values to remove (\code{"success"}, \code{"error"}, \code{"cancelled"}).}
}
\value{
Number of jobs removed.
}
\description{
Remove finished jobs from disk
}
\examples{
nm_job_cleanup()
}
