% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_list}
\alias{nm_job_list}
\title{List jobs in the job root}
\usage{
nm_job_list(
  job_root = nm_job_root(),
  local_only = FALSE,
  remote_only = FALSE,
  sync_active = TRUE
)
}
\arguments{
\item{job_root}{Job storage directory.}
}
\value{
A data.frame sorted by creation time (newest first).
}
\description{
List jobs in the job root
}
\examples{
nm_job_list()
}
