% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_log}
\alias{nm_job_log}
\title{Read worker log for a job}
\usage{
nm_job_log(job_id, tail = 40L, job_root = nm_job_root())
}
\arguments{
\item{job_id}{Job identifier.}

\item{tail}{Number of lines to return from the end of the log.}
}
\description{
Read worker log for a job
}
\examples{
# After nm_job_submit(), use nm_job_log(job$id)
}
