% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_result}
\alias{nm_job_result}
\title{Retrieve a completed job result}
\usage{
nm_job_result(job_id, job_root = nm_job_root())
}
\arguments{
\item{job_id}{Job identifier.}

\item{job_root}{Job storage directory.}
}
\value{
An \code{nm_fit} object.
}
\description{
Retrieve a completed job result
}
\examples{
# After a successful nm_job_submit(), use nm_job_result(job$id)
}
