% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job.R
\name{nm_job_watch_signature}
\alias{nm_job_watch_signature}
\title{Watch signature for job directory changes (for Shiny auto-refresh)}
\usage{
nm_job_watch_signature(job_root = nm_job_root())
}
\arguments{
\item{job_root}{Job storage directory.}
}
\value{
Character signature string.
}
\description{
Watch signature for job directory changes (for Shiny auto-refresh)
}
\examples{
nm_job_watch_signature()
}
