% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-data.R
\name{nm_max_scale_n}
\alias{nm_max_scale_n}
\title{Maximum number of compartment scaling factors (S1..Sn) supported}
\usage{
nm_max_scale_n()
}
\value{
Integer scalar (10, matching NONMEM $MODEL NCOMPS).
}
\description{
Maximum number of compartment scaling factors (S1..Sn) supported
}
\examples{
nm_max_scale_n()
}
