% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outcome.R
\name{nm_outcomes}
\alias{nm_outcomes}
\title{Combine endpoint declarations}
\usage{
nm_outcomes(...)
}
\arguments{
\item{...}{[nm_outcome()] objects or lists containing them.}
}
\value{
An `nm_outcomes` list.
}
\description{
Combine endpoint declarations
}
