% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-templates.R
\name{nm_piecewise}
\alias{nm_piecewise}
\title{Generate a piecewise-constant model expression}
\usage{
nm_piecewise(time = "TIME", knots, values)
}
\arguments{
\item{time}{Symbol containing time, normally `TIME`.}

\item{knots}{Increasing interval boundaries.}

\item{values}{One value/expression per interval (`length(knots) + 1`).}
}
\value{
Editable nested `ifelse()` code for `$PK/$PRED` or `$ERROR`.
}
\description{
Generate a piecewise-constant model expression
}
