% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{nm_project_delete}
\alias{nm_project_delete}
\title{Delete a modelling project}
\usage{
nm_project_delete(workspace, project)
}
\arguments{
\item{workspace}{Workspace object or path.}

\item{project}{Project id.}
}
\value{
`TRUE`, invisibly.
}
\description{
Delete a modelling project
}
