% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{nm_project_delete_snapshot}
\alias{nm_project_delete_snapshot}
\title{Delete a project snapshot}
\usage{
nm_project_delete_snapshot(workspace, project, snapshot)
}
\arguments{
\item{workspace}{Workspace object or path.}

\item{project}{Project id.}

\item{snapshot}{Snapshot id.}
}
\value{
`TRUE`, invisibly.
}
\description{
Delete a project snapshot
}
