% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{nm_project_load}
\alias{nm_project_load}
\title{Load a project snapshot}
\usage{
nm_project_load(workspace, project, snapshot = "latest")
}
\arguments{
\item{workspace}{Workspace object or path.}

\item{project}{Project id.}

\item{snapshot}{Snapshot id or `"latest"`.}
}
\value{
A serializable snapshot list.
}
\description{
Load a project snapshot
}
