% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{nm_project_load_diagnostics}
\alias{nm_project_load_diagnostics}
\title{Load saved diagnostics for a model run}
\usage{
nm_project_load_diagnostics(workspace, project, run)
}
\arguments{
\item{workspace}{Workspace object or path.}

\item{project}{Project id.}

\item{run}{Run id.}
}
\value{
A named list of saved diagnostic objects.
}
\description{
Load saved diagnostics for a model run
}
