% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-job-remote.R
\name{nm_remote_dataset_list}
\alias{nm_remote_dataset_list}
\title{List datasets on a remote cluster}
\usage{
nm_remote_dataset_list(server = NULL)
}
\arguments{
\item{server}{Server id or NULL for default.}
}
\value{
Data frame of datasets with id, label, md5.
}
\description{
List datasets on a remote cluster
}
