% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-remote-server.R
\name{nm_remote_server_list}
\alias{nm_remote_server_list}
\title{List configured remote scheduler servers}
\usage{
nm_remote_server_list()
}
\value{
Data frame with id, name, base_url, username (tokens omitted).
}
\description{
List configured remote scheduler servers
}
\examples{
nm_remote_server_list()
}
