% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-remote-server.R
\name{nm_remote_server_remove}
\alias{nm_remote_server_remove}
\title{Remove a remote server}
\usage{
nm_remote_server_remove(id)
}
\arguments{
\item{id}{Server id from \code{\link{nm_remote_server_list}}.}
}
\description{
Remove a remote server
}
