% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-remote-server.R
\name{nm_remote_server_set_default}
\alias{nm_remote_server_set_default}
\title{Set default remote server}
\usage{
nm_remote_server_set_default(id)
}
\arguments{
\item{id}{Server id.}
}
\description{
Set default remote server
}
