% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-workflow.R
\name{nm_report_design_save}
\alias{nm_report_design_save}
\title{Save a report workflow with a project}
\usage{
nm_report_design_save(workspace, project, design)
}
\arguments{
\item{workspace}{An [nm_workspace()] or path.}

\item{project}{Project id.}

\item{design}{An [nm_report_design()].}
}
\value{
The design id, invisibly.
}
\description{
Save a report workflow with a project
}
