% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-shiny-job-hub.R
\name{nm_shiny_job_hub_notify_local}
\alias{nm_shiny_job_hub_notify_local}
\title{Push an immediate job refresh to all connected Shiny sessions (local change).}
\usage{
nm_shiny_job_hub_notify_local(job_root = nm_job_root())
}
\arguments{
\item{job_root}{Job directory that changed.}
}
\description{
Push an immediate job refresh to all connected Shiny sessions (local change).
}
