% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-shiny-job-hub.R
\name{nm_shiny_job_hub_update}
\alias{nm_shiny_job_hub_update}
\title{Update hub subscription for a Shiny session.}
\usage{
nm_shiny_job_hub_update(
  session,
  job_root = NULL,
  remote_servers = NULL,
  sync_remote = NULL
)
}
\arguments{
\item{session}{Shiny session.}

\item{job_root}{Job directory to watch.}

\item{remote_servers}{Remote server ids to sync.}

\item{sync_remote}{If TRUE, refresh remote stubs in the hub worker.}
}
\description{
Update hub subscription for a Shiny session.
}
