% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gof.R
\name{nm_shrinkage}
\alias{nm_shrinkage}
\title{Shrinkage diagnostics for random effects}
\usage{
nm_shrinkage(fit)
}
\arguments{
\item{fit}{An \code{nm_fit} object.}
}
\description{
Shrinkage diagnostics for random effects
}
\examples{
\dontrun{
sim <- nm_synthetic_theo(n_sub = 2L, seed = 1L)
fit <- nm_est(sim$model, sim$data, method = "FO",
              control = list(maxit = 5L, compute_inference = FALSE))
nm_shrinkage(fit)
}
}
