% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{nm_sim_template_data}
\alias{nm_sim_template_data}
\title{Resolve simulation template data from design options}
\usage{
nm_sim_template_data(model, data, design = list())
}
\arguments{
\item{model}{Model object.}

\item{data}{Linked version dataset.}

\item{design}{List with design fields from the Shiny simulation dialog.}
}
\description{
Resolve simulation template data from design options
}
\examples{
sim <- nm_synthetic_theo(n_sub = 2L)
nm_sim_template_data(sim$model, sim$data)
}
