% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{nm_sim_workload}
\alias{nm_sim_workload}
\title{Estimate simulation workload (subjects x observation rows x replicates)}
\usage{
nm_sim_workload(model, data, n_sim = 1L)
}
\description{
Estimate simulation workload (subjects x observation rows x replicates)
}
\examples{
sim <- nm_synthetic_theo(n_sub = 2L)
nm_sim_workload(sim$model, sim$data, n_sim = 2L)
}
