% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-validation.R
\name{nm_support_matrix}
\alias{nm_support_matrix}
\title{Supported method / grad / engine matrix}
\usage{
nm_support_matrix()
}
\description{
Supported method / grad / engine matrix
}
\examples{
nm_support_matrix()
}
