% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-synthetic.R
\name{nm_synthetic_infusion}
\alias{nm_synthetic_infusion}
\title{Synthetic IV infusion 1-compartment dataset}
\usage{
nm_synthetic_infusion(n_sub = 10L, seed = 1L)
}
\arguments{
\item{n_sub}{Number of subjects.}

\item{seed}{Random seed.}
}
\description{
Synthetic IV infusion 1-compartment dataset
}
\examples{
nm_synthetic_infusion(n_sub = 2L)
}
