% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-synthetic.R
\name{nm_synthetic_pkpd_emax}
\alias{nm_synthetic_pkpd_emax}
\title{Synthetic PK/PD Emax-linked dataset}
\usage{
nm_synthetic_pkpd_emax(n_sub = 10L, seed = 1L)
}
\arguments{
\item{n_sub}{Number of subjects.}

\item{seed}{Random seed.}
}
\description{
Synthetic PK/PD Emax-linked dataset
}
\examples{
nm_synthetic_pkpd_emax(n_sub = 2L)
}
