% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-synthetic.R
\name{nm_synthetic_steady}
\alias{nm_synthetic_steady}
\title{Synthetic oral steady-state dataset}
\usage{
nm_synthetic_steady(n_sub = 10L, seed = 1L)
}
\arguments{
\item{n_sub}{Number of subjects.}

\item{seed}{Random seed.}
}
\description{
Synthetic oral steady-state dataset
}
\examples{
nm_synthetic_steady(n_sub = 2L)
}
