% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est.R
\name{nm_synthetic_theo}
\alias{nm_synthetic_theo}
\title{Synthetic THEO-like dataset for examples and tests}
\usage{
nm_synthetic_theo(n_sub = 10L, seed = 1L)
}
\arguments{
\item{n_sub}{Number of subjects.}

\item{seed}{Random seed.}
}
\value{
List with \code{model} and \code{data}.
}
\description{
Synthetic THEO-like dataset for examples and tests
}
\examples{
sim <- nm_synthetic_theo(n_sub = 2L, seed = 1L)
names(sim)
head(sim$data$data)
}
