% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-sim-ext.R
\name{nm_vpc_plot}
\alias{nm_vpc_plot}
\title{Plot a VPC / pcVPC (requires ggplot2)}
\usage{
nm_vpc_plot(x, ...)
}
\arguments{
\item{x}{An \code{nm_vpc} object from \code{\link{nm_vpc}}.}

\item{...}{Unused.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a VPC / pcVPC (requires ggplot2)
}
\examples{
\dontrun{
nm_vpc_plot(nm_vpc(fit, n_sim = 100L))
}
}
