% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_delete_project}
\alias{nm_workspace_delete_project}
\title{Delete a project from the workspace}
\usage{
nm_workspace_delete_project(project, root = nm_workspace_root())
}
\description{
Delete a project from the workspace
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("tmp", path = ws, template = "empty")
nm_workspace_delete_project("tmp", root = ws)
}
}
