% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_delete_run}
\alias{nm_workspace_delete_run}
\title{Delete an estimation run}
\usage{
nm_workspace_delete_run(
  project,
  version_id,
  run_id,
  root = nm_workspace_root()
)
}
\description{
Delete an estimation run
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("demo", path = ws, template = "theo")
ver <- nm_workspace_new_version("demo", root = ws)
run_id <- nm_workspace_new_run_id("demo", ver, root = ws)
nm_workspace_delete_run("demo", ver, run_id, root = ws)
}
}
