% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_delete_version}
\alias{nm_workspace_delete_version}
\title{Delete a model version and its estimation runs}
\usage{
nm_workspace_delete_version(project, version_id, root = nm_workspace_root())
}
\description{
Delete a model version and its estimation runs
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("demo", path = ws, template = "theo")
ver <- nm_workspace_new_version("demo", root = ws)
nm_workspace_delete_version("demo", ver, root = ws)
}
}
