% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_init}
\alias{nm_workspace_init}
\title{Initialize a workspace directory}
\usage{
nm_workspace_init(path = nm_workspace_root(), create_demo_project = TRUE)
}
\arguments{
\item{path}{Directory to use as workspace root.}

\item{create_demo_project}{Create a THEO demo project when \code{TRUE}.}
}
\value{
Invisibly, the workspace root path.
}
\description{
Initialize a workspace directory
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
}
}
