% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_list_datasets}
\alias{nm_workspace_list_datasets}
\title{List dataset files in a project}
\usage{
nm_workspace_list_datasets(project, root = nm_workspace_root())
}
\arguments{
\item{project}{Project name.}

\item{root}{Workspace root.}
}
\value{
Data frame with \code{file} (basename) and \code{path} (relative to project).
}
\description{
List dataset files in a project
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("theo_demo", path = ws, template = "theo")
nm_workspace_list_datasets("theo_demo", root = ws)
}
}
