% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_list_project_sims}
\alias{nm_workspace_list_project_sims}
\title{List all simulations in a project (across model versions)}
\usage{
nm_workspace_list_project_sims(project, root = nm_workspace_root())
}
\description{
List all simulations in a project (across model versions)
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("theo_demo", path = ws, template = "theo")
nm_workspace_list_project_sims("theo_demo", root = ws)
}
}
