% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_list_projects}
\alias{nm_workspace_list_projects}
\title{List projects in the workspace}
\usage{
nm_workspace_list_projects(root = nm_workspace_root())
}
\arguments{
\item{root}{Workspace root.}
}
\value{
Character vector of project names.
}
\description{
List projects in the workspace
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_list_projects(ws)
}
}
