% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_list_versions}
\alias{nm_workspace_list_versions}
\alias{nm_workspace_list_models}
\title{List model versions in a project}
\usage{
nm_workspace_list_versions(project, root = nm_workspace_root())

nm_workspace_list_models(project, root = nm_workspace_root())
}
\arguments{
\item{project}{Project name.}

\item{root}{Workspace root.}
}
\value{
Character vector of version ids.
}
\description{
Model versions are distinct control streams (different model code) within a project.
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("theo_demo", path = ws, template = "theo")
nm_workspace_list_versions("theo_demo", root = ws)
}
}
