% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_load_run_meta}
\alias{nm_workspace_load_run_meta}
\title{Load estimation run metadata}
\usage{
nm_workspace_load_run_meta(
  project,
  version_id,
  run_id,
  root = nm_workspace_root()
)
}
\arguments{
\item{project}{Project name.}

\item{version_id}{Model version id.}

\item{run_id}{Estimation run id.}

\item{root}{Workspace root.}
}
\value{
Metadata list from \code{meta.json}, or \code{NULL}.
}
\description{
Load estimation run metadata
}
