% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_model_path}
\alias{nm_workspace_model_path}
\title{Path to a model version control file}
\usage{
nm_workspace_model_path(project, version_id, root = nm_workspace_root())
}
\arguments{
\item{project}{Project name.}

\item{version_id}{Model version id.}
}
\description{
Path to a model version control file
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("theo_demo", path = ws, template = "theo")
ver <- nm_workspace_list_versions("theo_demo", root = ws)[[1L]]
nm_workspace_model_path("theo_demo", ver, root = ws)
}
}
