% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_new_sim_id}
\alias{nm_workspace_new_sim_id}
\title{Allocate a new simulation id for a model version}
\usage{
nm_workspace_new_sim_id(project, version_id, root = nm_workspace_root())
}
\description{
Allocate a new simulation id for a model version
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("theo_demo", path = ws, template = "theo")
ver <- nm_workspace_list_versions("theo_demo", root = ws)[[1L]]
nm_workspace_new_sim_id("theo_demo", ver, root = ws)
}
}
