% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-workspace.R
\name{nm_workspace_reports_dir}
\alias{nm_workspace_reports_dir}
\title{Reports directory for a project (confined to workspace)}
\usage{
nm_workspace_reports_dir(project, root = nm_workspace_root())
}
\description{
Reports directory for a project (confined to workspace)
}
\examples{
\dontrun{
ws <- tempfile()
nm_workspace_init(ws, create_demo_project = FALSE)
nm_workspace_create_project("demo", path = ws, template = "empty")
nm_workspace_reports_dir("demo", root = ws)
}
}
