% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-est.R
\name{print.nm_fit}
\alias{print.nm_fit}
\title{Print method for \code{nm_fit} objects}
\usage{
\method{print}{nm_fit}(x, ...)
}
\arguments{
\item{x}{An \code{nm_fit} object.}

\item{...}{Unused.}
}
\description{
Print method for \code{nm_fit} objects
}
\examples{
\dontrun{
sim <- nm_synthetic_theo(n_sub = 2L, seed = 1L)
fit <- nm_est(sim$model, sim$data, method = "FO",
              control = list(maxit = 5L, compute_inference = FALSE))
print(fit)
}
}
