% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{print.nm_nca}
\alias{print.nm_nca}
\title{Print an NCA result}
\usage{
\method{print}{nm_nca}(x, ...)
}
\arguments{
\item{x}{An \code{nm_nca} object.}
\item{...}{Unused.}
}
\description{Print a compact noncompartmental analysis summary.}
