% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-bayes.R
\name{print.summary.nm_fit_bayes}
\alias{print.summary.nm_fit_bayes}
\title{Print method for Bayesian MCMC summary objects}
\usage{
\method{print}{summary.nm_fit_bayes}(x, ...)
}
\arguments{
\item{x}{A \code{summary.nm_fit_bayes} object.}

\item{...}{Unused.}
}
\description{
Print method for Bayesian MCMC summary objects
}
\examples{
\dontrun{
sim <- nm_synthetic_theo(n_sub = 2L, seed = 1L)
fit <- nm_est(sim$model, sim$data, method = "BAYES", n_burn = 5L, n_sample = 10L,
              control = list(compute_inference = FALSE))
print(summary(fit))
}
}
