% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-profiling.R
\name{print.nm_profile}
\alias{print.nm_profile}
\alias{profile.nm_fit}
\title{Summarise profiling data attached to a fit}
\usage{
\method{print}{nm_profile}(x, ...)

\method{profile}{nm_fit}(object, ...)
}
\arguments{
\item{x}{A profiling summary data frame.}

\item{...}{Unused.}

\item{object}{An \code{nm_fit} object.}
}
\description{
Summarise profiling data attached to a fit
}
\examples{
\dontrun{
options(LibeRation.profile = TRUE)
sim <- nm_synthetic_theo(n_sub = 2L)
fit <- nm_est(sim$model, sim$data, method = "FO",
              control = list(maxit = 3L, compute_inference = FALSE))
options(LibeRation.profile = FALSE)
if (!is.null(fit$profile)) print(profile(fit))
}
\dontrun{
options(LibeRation.profile = TRUE)
sim <- nm_synthetic_theo(n_sub = 2L)
fit <- nm_est(sim$model, sim$data, method = "FO",
              control = list(maxit = 3L, compute_inference = FALSE))
options(LibeRation.profile = FALSE)
profile(fit)
}
}
