% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{renderLiberWorkbench}
\alias{renderLiberWorkbench}
\title{Render a LibeR workbench in Shiny}
\usage{
renderLiberWorkbench(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Expression returning [liber_workbench()].}

\item{env}{Evaluation environment.}

\item{quoted}{Whether `expr` is quoted.}
}
\description{
Render a LibeR workbench in Shiny
}
