% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-gof.R
\name{residuals.nm_fit}
\alias{residuals.nm_fit}
\title{Extract residuals from a fit}
\usage{
\method{residuals}{nm_fit}(object, ...)
}
\arguments{
\item{object}{An \code{nm_fit} object.}
}
\description{
Extract residuals from a fit
}
\examples{
\dontrun{
sim <- nm_synthetic_theo(n_sub = 2L, seed = 1L)
fit <- nm_est(sim$model, sim$data, method = "FO",
              control = list(maxit = 5L, compute_inference = FALSE))
head(residuals(fit))
}
}
