% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm-fo.R
\name{.nm_fo_report_objective}
\alias{.nm_fo_report_objective}
\title{FO omega step: \eqn{\sum_i \eta_i^{\top} \Omega^{-1} \eta_i + \log|\Omega|} with fixed etas}
\usage{
.nm_fo_report_objective(model, data, par, eta_mat, pk_engine = "auto")
}
\description{
FO omega step: \eqn{\sum_i \eta_i^{\top} \Omega^{-1} \eta_i + \log|\Omega|} with fixed etas
}
\keyword{internal}
