% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{nm_etab}
\alias{nm_etab}
\title{Empirical Bayes estimates and shrinkage}
\usage{
nm_etab(fit)
}
\arguments{
\item{fit}{An `nm_fit`.}
}
\value{
A list with subject ETA table and component-wise shrinkage.
}
\description{
Empirical Bayes estimates and shrinkage
}
