% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-templates.R
\name{nm_structural_templates}
\alias{nm_structural_templates}
\title{Catalogue of advanced structural model templates}
\usage{
nm_structural_templates()
}
\value{
A data frame describing templates accepted by
  [nm_model_template()].
}
\description{
Catalogue of advanced structural model templates
}
