% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nca.R
\name{print.nm_nca}
\alias{print.nm_nca}
\title{Print an NCA result}
\usage{
\method{print}{nm_nca}(x, ...)
}
\arguments{
\item{x}{An `nm_nca` object.}

\item{...}{Unused.}
}
\value{
`x`, invisibly.
}
\description{
Print an NCA result
}
